Fix for MPLY-7558
https://eegeo-team.atlassian.net/browse/MPLY-7558
Added 3 events to the WebGL build, TransitionStart, TransitionEnd and
Zoom. The transition events are fired when SetView or SetViewToBounds
are called to jump/transition the camera to a new location. This bug
could have been fixed by hiding/showing markers behind/infront of the
world on the event moveend but this is fired after all camera movements.
Buddy: Vimmy