Fix visibility of leaflet polygons in indoor maps
The logic was pretty much correct, but the visibility was being
clobbered by onPan / onZoom / transitionend, as they trigger a function
call to _hideMarkersBehindEarth.

This function is misleadingly named; it should actually be called
"_updateVisibilityForAllLayers" or similar, as it iterates over all
features, not just markers.

Additionally: Fix an edge case where, if the user creates a polygon with
a floor index of 0, it would've triggered the wrong logic (as !0
resolves to true). Fixed by comparing to === undefined, instead.

Buddy: Ben P