Enter indoors if we have config.latlng & indoorId
Previously, we gated access to indoor maps based on whether an indoor
marker was present. In certain rare cases, we need to be able to enter
an indoor map that does not have a corresponding marker (e.g.
MPLY-11248)

If the user has supplied a latLng via configuration such as query
string, we can try using this latLng rather than the entrance marker
latLng.

If neither is provided, just bail as before.

Buddy: Malc B