Correct misnamed unused parameter.
All other API objects constructed in the same palce take the
emscriptenMemory object as that parameter. This is what the function is
being passed at its call site. However, the parameter name is
emscriptenModule, which has upset the linter now we're actually passing
that object, too.

This doesn't change any *behaviour* - the argument still has the same
value. It just makes the name match the type.