Update tests to be compatible with newer Emscripten versions
Several changes to Emscripten mandated changes to the tests:
* Emscripten's synchronous loading now takes less time than its
asynchronous loading, so waiting for the module to be loaded is now
needed. We should have been doing this already, but we'd just got lucky
and never needed it before.
* Emscripten no longer gives error codes plucked from nowhere when
throwing errors. If given a useful message, it provides that, and
when no message is defined, it gives `undefined` as the reason. This
means that the whitelisted errors needed changing. It also means that
the errors would be logged during the tests instead of logging
meaningless nummbers, so a message was added to assure people looking at
test results that this is normal.
* Under certain circumstances, benign errors detected by Emscripten
weren't being handled until after the tests causing them were finished,
killing Jasmine at some indeterminate point in the future, so a custom
rejected promise handler is needed insstead of Emscripten's.
* The Emscripten Module object wasn't being fully reset between tests -
the event handlers it was registering with Node were persisting, and
actually causing Node to reject later ones oon the grounds that there
was a leak.
* One test was passing the wrong number of arguments to a function due
to a change in this PR
https://github.com/wrld3d/wrld.js/commit/4a3976e826be00f35d8610b527e3661c6eb17075
not being applied to the tests.

Other changes to the Emscripten module were necessary to get this
working, but are in another repo.