Make tests compatible with minified/mangled names
If, as is sensible for testing, you were to run the tests using an
unminified, unmangled verison of the SDK, it would say that
FUNCTION_TABLE[something] is not a function, and that was whitelisted as
an expected error. However, mangling and minification can change the
name of the function table, so now we check for any valid JavaScript
variable name.