Download memory initialiser when testing
Emscripten now requires the memory initialiser to be present when
running the tests.

Use curl --compressed instead of manually unzipping as this also handles
uncompressed files if we ever decide we don't want to host the memory
initialiser gzipped any more, or want to change to a more efficient
format.