Ensure Unicode strings aren't truncated.
In UTF-8, code points beyond 127 require multiple bytes to express.
According to Jonty, we use Unicode strings (as we have to support things
like place names in a range of alphabets).

Before this commit, we were mallocing a buffer of exactly 32 bytes in
one function, or as many bytes as the string had code points in another,
and the correct length in a third.

Prior to the Emscripten upgrade, when strings exceeded this length, the
32-byte function would truncate them to 32 bytes worth of characters
(which could possibly be part way through the representation of a code
point) and the other function would run past the end of the buffer,
corrupting memeory, but potentially making the whole string accessible
to something that consumed it.

After the initial part of the Emscripten upgrade, the function that
could corrupt memeory switched to truncating the string (again,
potentially part way through a character).

This commit switches to querying the actual number of bytes required for
the string using an Emscripten-provided function. Strings should pass
through the JS/Emscripten boundary intact.