Emscripten upgrade [Don't merge yet] (#70)
* Replace runtime with emscriptenModule

Runtime has been removed. Its member functions (such as addFunction) have been moved to Module.
Therefore we must access such functions via the module.

* Change Pointer_stringify to UTF8ToString

Buddy Stefano

* Correct misnamed unused parameter.

All other API objects constructed in the same palce take the
emscriptenMemory object as that parameter. This is what the function is
being passed at its call site. However, the parameter name is
emscriptenModule, which has upset the linter now we're actually passing
that object, too.

This doesn't change any *behaviour* - the argument still has the same
value. It just makes the name match the type.

* Switch from writeStringToMemory to stringToUTF8.

writeStringToMemory is deprecated, so needs replacing.
stringToUTF8 needs an extra argument for the maximum length to write to
the buffer. This is given the same value as malloc is.

Buddy Stefano

* Rename runtime argument of Precache API to emscriptenModule

This is to reflect that Emscripten have removed Module.Runtime and moved
its fields and functions into Module. This file was missed in the first
pass of renamings.

Buddy Stefano

* Replace use of Module.Runtime in tests.

Module.Runtime was removed from Emscripten and its fields and functions
were moved to Module.

Buddy Stefano

* Update tests to be compatible with newer Emscripten versions

Several changes to Emscripten mandated changes to the tests:
* Emscripten's synchronous loading now takes less time than its
asynchronous loading, so waiting for the module to be loaded is now
needed. We should have been doing this already, but we'd just got lucky
and never needed it before.
* Emscripten no longer gives error codes plucked from nowhere when
throwing errors. If given a useful message, it provides that, and
when no message is defined, it gives `undefined` as the reason. This
means that the whitelisted errors needed changing. It also means that
the errors would be logged during the tests instead of logging
meaningless nummbers, so a message was added to assure people looking at
test results that this is normal.
* Under certain circumstances, benign errors detected by Emscripten
weren't being handled until after the tests causing them were finished,
killing Jasmine at some indeterminate point in the future, so a custom
rejected promise handler is needed insstead of Emscripten's.
* The Emscripten Module object wasn't being fully reset between tests -
the event handlers it was registering with Node were persisting, and
actually causing Node to reject later ones oon the grounds that there
was a leak.
* One test was passing the wrong number of arguments to a function due
to a change in this PR
https://github.com/wrld3d/wrld.js/commit/4a3976e826be00f35d8610b527e3661c6eb17075
not being applied to the tests.

Other changes to the Emscripten module were necessary to get this
working, but are in another repo.

* Make tests compatible with minified/mangled names

If, as is sensible for testing, you were to run the tests using an
unminified, unmangled verison of the SDK, it would say that
FUNCTION_TABLE[something] is not a function, and that was whitelisted as
an expected error. However, mangling and minification can change the
name of the function table, so now we check for any valid JavaScript
variable name.

* Download memory initialiser when testing

Emscripten now requires the memory initialiser to be present when
running the tests.

Use curl --compressed instead of manually unzipping as this also handles
uncompressed files if we ever decide we don't want to host the memory
initialiser gzipped any more, or want to change to a more efficient
format.

* Manually check gzip compression for mem file

It is not possible to use the "--compressed" flag with curl, because the
version shipped with Windows doesn't have libz built in with it.

Buddy: Chris

* Ensure Unicode strings aren't truncated.

In UTF-8, code points beyond 127 require multiple bytes to express.
According to Jonty, we use Unicode strings (as we have to support things
like place names in a range of alphabets).

Before this commit, we were mallocing a buffer of exactly 32 bytes in
one function, or as many bytes as the string had code points in another,
and the correct length in a third.

Prior to the Emscripten upgrade, when strings exceeded this length, the
32-byte function would truncate them to 32 bytes worth of characters
(which could possibly be part way through the representation of a code
point) and the other function would run past the end of the buffer,
corrupting memeory, but potentially making the whole string accessible
to something that consumed it.

After the initial part of the Emscripten upgrade, the function that
could corrupt memeory switched to truncating the string (again,
potentially part way through a character).

This commit switches to querying the actual number of bytes required for
the string using an Emscripten-provided function. Strings should pass
through the JS/Emscripten boundary intact.


Overall buddies: Stefano & Mark