Add hardware accleration checking function
Adds a function to the public API that can check if the browser supports
hardware-accelerated WebGL on the current machine.

This creates a new context with the failIfMajorPerformanceCaveat flag
enabled, which will fail in the absence of hardware acceleration. This
can't be queried from the existing coontext without a WebGL extension
that isn't ubiquitous, so there's no other viable approach.

Buddy Mark