Don't set background colour twice
Previously, we were setting the draw clear colour both as the WebGL
clear colour and the background colour of the HTML canvas. In the case
where it was translucent, the colour contributed twice what it was
supposed to, and in the case where it was opaque, you couldn't see the
canvas' background colour anyway, so it was pointless.

Rendering the map onto a transparent background isn't actually
supported, so, bugs notwithstanding, we always hit the pointless case.