Add getFloorZOrder to IndoorMapFloor
I noticed that the type definitions were missing _getFloorId, which is to be expected because it is a private method.

On the other hand, there was no other way of getting the equivalent with a public method.

I have added getFloorZOrder, which does what it says on the tin, marked the other two methods as deprecated pointing towards their more aptly named versions.