Explicitly throw exception when using native polygon with L.Map
The unit tests expected it to be thrown immediately, but with the change
to make native polygons into Leaflet layers, it doesn't get added to the
map until later, deferring throwing the exception until then.

If we throw it explicitly in the beforeAdd function, which is called
synchronously, the tests pass, and also it's a little clearer where a
mistake happened as the stack trace will make more sense.